[SYNC] NCCL-Tests v2.16.7 (#147)
* Print the name of the program being executed
One thing missing from the stdout of each performance test is
the name of the test that is actually being run.

This patch adds 2 new messages to the stdout. At the beginning
of the execution of a test (e.g. sendrecv_perf) we will now
see this message:

  Collective test starting: sendrecv_perf

And at the end, we will now see this:

  Collective test concluded: sendrecv_perf

This is needed when running several tests consecutively and we're
trying to parse the stdout to collect the results.

For example, using a Python script to parse the stdout, one could
retrieve the results for each test and plot them on a graph. This
patch makes it easier to implement such a script.

Signed-off-by: Martin Belanger <martin.belanger@dell.com>

* Minor fix to Makefile
Move comments to separate lines

* Add extra reserved space during maxBytes calculation
Also, don't allow minBytes > maxBytes

---------

Signed-off-by: Martin Belanger <martin.belanger@dell.com>
Co-authored-by: Martin Belanger <martin.belanger@dell.com>
Co-authored-by: David Addison <daddison@nvidia.com>
Co-authored-by: David Addison <2538330+AddyLaddy@users.noreply.github.com>